refactor(jco): move type generation functions to separate file
This commit separates out type generation to another file from the
transpilation logic, but keeps the exports that were previously part
of transpile in place to ensure backwards compatibility.
