+++ packages/jco/src/common.d.ts
@@ -10,6 +10,7 @@ export function table(data: any, align?: any[]): string;
  */
 export function getTmpDir(): Promise<string>;
 export function spawnIOTmp(cmd: any, input: any, args: any): Promise<Buffer<ArrayBufferLike>>;
+export function writeFiles(files: any, summaryTitle: any): Promise<void>;
 export const isWindows: boolean;
 export { readFileCli as readFile };
 declare function readFileCli(file: any, encoding: any): Promise<Buffer<ArrayBufferLike>>;
